# backend/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API REST para gestão de empresas, pesquisas, dashboards, auditoria e submissões anônimas.\n\nFluxo sugerido para clientes frontend:\n1) Autenticar em /auth/login\n2) Enviar token JWT no header Authorization: Bearer \u003ctoken\u003e\n3) Consumir endpoints protegidos em /api/v1\n",
        "title": "API de Pesquisa de Clima Organizacional",
        "termsOfService": "https://example.com/terms",
        "contact": {
            "name": "Equipe Backend",
            "email": "backend@example.com"
        },
        "license": {
            "name": "Proprietary"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/v1/auth/change-password": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Valida a senha atual e altera para a nova senha informada.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Alterar senha do usuário autenticado",
                "parameters": [
                    {
                        "description": "Senha atual e nova senha",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_infrastructure_auth.ChangePasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto_response.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto_response.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto_response.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto_response.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/forgot-password": {
            "post": {
                "description": "Inicia o fluxo de recuperação de senha para o email informado.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Solicitar recuperação de senha",
                "parameters": [
                    {
                        "description": "Email para recuperação",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_infrastructure_auth.ForgotPasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto_response.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto_response.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto_response.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/login": {
            "post": {
                "description": "Realiza login com email e senha e retorna token JWT para uso nos endpoints protegidos.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Autenticar administrador",
                "parameters": [
                    {
                        "description": "Credenciais de acesso",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_infrastructure_auth.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto_response.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto_response.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto_response.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto_response.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/logout": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Registra logout do usuário autenticado.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Encerrar sessão",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto_response.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto_response.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/refresh": {
            "post": {
                "description": "Valida o token atual e emite um novo token JWT.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Renovar token JWT",
                "parameters": [
                    {
                        "description": "Token atual",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_infrastructure_auth.RefreshTokenRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto_response.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto_response.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto_response.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto_response.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/validate": {
            "post": {
                "description": "Valida um token JWT informado e retorna dados básicos do usuário associado.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Validar token JWT",
                "parameters": [
                    {
                        "description": "Token para validação",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_infrastructure_auth.ValidateTokenRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto_response.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto_response.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto_response.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto_response.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_infrastructure_auth.ChangePasswordRequest": {
            "type": "object",
            "required": [
                "nova_senha",
                "senha_atual"
            ],
            "properties": {
                "nova_senha": {
                    "description": "Nova senha",
                    "type": "string",
                    "maxLength": 128,
                    "minLength": 8,
                    "example": "NovaSenha@123"
                },
                "senha_atual": {
                    "description": "Senha atual",
                    "type": "string",
                    "example": "SenhaAtual@123"
                }
            }
        },
        "internal_infrastructure_auth.ForgotPasswordRequest": {
            "type": "object",
            "required": [
                "email"
            ],
            "properties": {
                "email": {
                    "description": "Email para recuperação",
                    "type": "string",
                    "example": "admin@empresa.com"
                }
            }
        },
        "internal_infrastructure_auth.LoginRequest": {
            "type": "object",
            "required": [
                "email",
                "senha"
            ],
            "properties": {
                "email": {
                    "description": "Email do usuário",
                    "type": "string",
                    "example": "admin@empresa.com"
                },
                "senha": {
                    "description": "Senha em texto plano",
                    "type": "string",
                    "example": "SenhaForte@123"
                }
            }
        },
        "internal_infrastructure_auth.RefreshTokenRequest": {
            "type": "object",
            "required": [
                "token"
            ],
            "properties": {
                "token": {
                    "description": "Token JWT atual",
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                }
            }
        },
        "internal_infrastructure_auth.ValidateTokenRequest": {
            "type": "object",
            "required": [
                "token"
            ],
            "properties": {
                "token": {
                    "description": "Token JWT a ser validado",
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                }
            }
        },
...
        "organizational-climate-survey_backend_internal_application_dto_response.APIResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Codigo de erro para consumo de clientes",
                    "type": "string",
                    "example": "BAD_REQUEST"
                },
                "data": {
                    "description": "Dados retornados, genéricos"
                },
                "error": {
                    "description": "Mensagem de erro, se houver",
                    "type": "string",
                    "example": "Parâmetro inválido"
                },
                "message": {
                    "description": "Mensagem opcional para o usuário",
                    "type": "string",
                    "example": "Operação realizada com sucesso"
                },
                "success": {
                    "description": "Indica se a operação foi bem-sucedida",
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}